fashion-vision-iit/requirements.txt
streamlit
tensorflow
keras
opencv-python
Pillow
numpy
colorthief
